Moved files around to prepare CLI execution
