+++ src/common/SaveFile.ts
@@ -1,10 +1,10 @@
 import { createHash } from 'crypto';
 import log from 'electron-log';
 import { readFile, writeFile } from 'fs/promises';
-import { Edge, Node, Viewport } from 'reactflow';
 import { EdgeData, FileOpenResult, NodeData, Version } from './common-types';
 import { currentMigration, migrate } from './migrations';
 import { versionGt } from './version';
+import type { Edge, Node, Viewport } from 'reactflow';
 
 export interface SaveData {
     nodes: Node<NodeData>[];
